credence-pi #27 stage 3: the body emits taint features live (SafetyTracker) — v0.2.0
The plugin now computes the safety/taint-flow features so live OpenClaw sessions exercise
multi-outcome governance. The body is the single source of truth for the taint computation
(the eval re-exports from it, mirroring FeatureTracker), so the offline-trained harm posterior
and the live signal agree.

- openclaw-plugin/src/safety.ts: the taint primitives (moved from eval/adapters/risk.ts,
  verbatim) + a SafetyTracker holding per-session causal taint state. observeResult()
  accumulates distinctive tokens + imperative verbs from a tool RESULT (the untrusted source
  the agent consumes); extractSafety() emits the 4 features (action-class, taint-flow,
  injected-imperative, cred-exfil-chain) for a proposed call against tokens seen BEFORE it.
- index.ts: beforeToolCall merges the safety features into the sensor event (additive —
  the daemon reads only its declared subsets, so any daemon version is unaffected);
  afterToolCall feeds the tool result into observeResult.
- eval/adapters/risk.ts: now re-exports the primitives from the plugin (single source).
  Re-extracting ATBench produces BYTE-IDENTICAL events ⇒ the shipped harm posterior stays valid.

Verified: plugin builds (tsc); 35 plugin tests pass (5 new SafetyTracker tests — a tool result
taints a later sink, the cred→external chain fires, clean sessions stay untainted, sessions are
isolated); ATBench re-extract identical; credence-lint clean.

This completes the multi-outcome CODE (brain + harm posterior + body). Remaining is operational
only: npm publish of v0.2.0 (needs Guy's 2FA OTP) + the daemon image (CI-published on master).
Harm governance ships ON in safe ask-mode (confirm, never silent-block).
